Re-organize prometheus tests into the unit testing category
